Create new journey for parent requests
