+++ src/app/core/services/breadcrumb.service.ts
@@ -36,9 +36,8 @@ import {
   staffRecordsTabJourney,
   trainingAndQualificationsTabJourney,
   workplaceTabJourney,
-  becomeAParentJourney,
-  linkToParentJourney,
 } from '@core/breadcrumb/journey.workplaces';
+import { becomeAParentJourney, linkToParentJourney } from '@core/breadcrumb/journey.parent-requests';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { distinctUntilChanged, filter, map } from 'rxjs/operators';
 import { parse } from 'url';
